+++ app/providers/connections/ble.ts
@@ -1,10 +1,10 @@
-import { Injectable, NgZone } from '@angular/core';
+import { NgZone } from '@angular/core';
 
 import { Subject } from 'rxjs/Subject';
 import 'rxjs/add/operator/map';
 
-import { Connection, Device, ConnectionProvider } from './connection';
-import { Logger, Plugins } from '../providers';
+import { Connection } from '../cu';
+import { Logger } from '../logger';
 
 const SERVICE_UUID = '39df7777-b1b4-b90b-57f1-7144ae4e4a6a';
 const OUTPUT_UUID = '39df8888-b1b4-b90b-57f1-7144ae4e4a6a';
@@ -13,13 +13,12 @@ const NOTIFY_UUID = '39df9999-b1b4-b90b-57f1-7144ae4e4a6a';
 const DOLLAR = '$'.charCodeAt(0);
 
 export class BLEConnection extends Subject<ArrayBuffer> implements Connection {
-
-  private logEnabled = false;
   
-  constructor(private ble: any, private logger: Logger, private zone: NgZone, private peripheral: any) {
+  constructor(private ble: any, private peripheral: any, private zone: NgZone, private logger?: Logger) {
     super();
     this.log('Creating BLE connection');
-    this.ble.startNotification(this.peripheral.id, SERVICE_UUID, NOTIFY_UUID,
+    this.ble.startNotification(
+      this.peripheral.id, SERVICE_UUID, NOTIFY_UUID,
       data => {
         this.log('Received BLE data', data);
         // strip trailing '$' and prepend missing '0'/'?' for notifications
@@ -31,7 +30,6 @@ export class BLEConnection extends Subject<ArrayBuffer> implements Connection {
         this.zone.run(() => this.next(view.buffer)); 
       },
       error => {
-        this.logger.error("BLE receive error", error);
         this.zone.run(() => this.error(error));
       }
     );
@@ -60,7 +58,7 @@ export class BLEConnection extends Subject<ArrayBuffer> implements Connection {
   }
   
   private log(message: string, data?: ArrayBuffer) {
-    if (this.logEnabled) {
+    if (this.logger) {
       let now = Date.now();
       if (data) {
         this.logger.debug(now.toString() + ": " + message + " " + String.fromCharCode.apply(null, new Uint8Array(data)));
@@ -69,31 +67,14 @@ export class BLEConnection extends Subject<ArrayBuffer> implements Connection {
       }
     }
   }
-}
 
-@Injectable()
-export class BLEProvider extends ConnectionProvider {
-  constructor(private logger: Logger, private plugins: Plugins, private zone: NgZone) {
-    super();
-  }   
-
-  connect(device: Device) {
-    return this.enable().then(ble => {
-      return new Promise<Connection>((resolve, reject) => {
-        ble.connect(
-          device.id, 
-          peripheral => resolve(new BLEConnection(ble, this.logger, this.zone, peripheral)), 
-          reject
-        );
-      });
-    });
-  }
-
-  private enable(): Promise<any> {
-    return this.plugins.get('ble').then(ble => {
-      return new Promise((resolve, reject) => {
-        ble.enable(() => resolve(ble), () => reject(new Error('BLE not enabled')));
-      });
+  static connect(ble: any, id: string, zone: NgZone, logger: Logger) {
+    return new Promise<Connection>((resolve, reject) => {
+      ble.connect(
+        id, 
+        peripheral => resolve(new BLEConnection(ble, peripheral, zone, logger)), 
+        reject
+      );
     });
   }
 }