+++ app/providers/connections/index.ts
@@ -1,0 +1,3 @@
+export * from './ble';
+export * from './demo';
+export * from './serial';