+++ app/providers/index.ts
@@ -1,4 +1,5 @@
 export * from './cu';
+export * from './devices';
 export * from './logger';
 export * from './plugins';
 export * from './race-control';